--- a/processor/deltatocumulativeprocessor/internal/delta/delta.go
+++ b/processor/deltatocumulativeprocessor/internal/delta/delta.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 
 	"go.opentelemetry.io/collector/pdata/pcommon"
+	"go.opentelemetry.io/collector/pdata/pmetric"
 
 	exp "github.com/open-telemetry/opentelemetry-collector-contrib/internal/exp/metrics/streams"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/deltatocumulativeprocessor/internal/data"
@@ -83,10 +84,17 @@ func (e ErrGap) Error() string {
 	return fmt.Sprintf("gap in stream from %s to %s. samples were likely lost in transit", e.From, e.To)
 }
 
+type Type interface {
+	pmetric.NumberDataPoint | pmetric.HistogramDataPoint | pmetric.ExponentialHistogramDataPoint
+
+	StartTimestamp() pcommon.Timestamp
+	Timestamp() pcommon.Timestamp
+}
+
 // AccumulateInto adds state and dp, storing the result in state
 //
 //	state = state + dp
-func AccumulateInto[P data.Point[P]](state P, dp P) error {
+func AccumulateInto[T Type](state, dp T) error {
 	switch {
 	case dp.StartTimestamp() < state.StartTimestamp():
 		// belongs to older series
@@ -96,6 +104,16 @@ func AccumulateInto[P data.Point[P]](state P, dp P) error {
 		return ErrOutOfOrder{Last: state.Timestamp(), Sample: dp.Timestamp()}
 	}
 
-	state.Add(dp)
+	switch dp := any(dp).(type) {
+	case pmetric.NumberDataPoint:
+		state := any(state).(pmetric.NumberDataPoint)
+		data.Number{NumberDataPoint: state}.Add(data.Number{NumberDataPoint: dp})
+	case pmetric.HistogramDataPoint:
+		state := any(state).(pmetric.HistogramDataPoint)
+		data.Histogram{HistogramDataPoint: state}.Add(data.Histogram{HistogramDataPoint: dp})
+	case pmetric.ExponentialHistogramDataPoint:
+		state := any(state).(pmetric.ExponentialHistogramDataPoint)
+		data.ExpHistogram{DataPoint: state}.Add(data.ExpHistogram{DataPoint: dp})
+	}
 	return nil
 }