--- a/processor/deltatocumulativeprocessor/internal/lineartelemetry/attr.go
+++ b/processor/deltatocumulativeprocessor/internal/lineartelemetry/attr.go
@@ -1,4 +1,7 @@
-package telemetry
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package telemetry // import "github.com/open-telemetry/opentelemetry-collector-contrib/processor/deltatocumulativeprocessor/internal/lineartelemetry"
 
 import "go.opentelemetry.io/otel/attribute"