--- a/processor/deltatocumulativeprocessor/internal/lineartelemetry/attr.go
+++ b/processor/deltatocumulativeprocessor/internal/lineartelemetry/attr.go
@@ -1,0 +1,9 @@
+package telemetry
+
+import "go.opentelemetry.io/otel/attribute"
+
+type Attributes []attribute.KeyValue
+
+func (a *Attributes) Set(attr attribute.KeyValue) {
+	*a = append(*a, attr)
+}