--- a/processor/deltatocumulativeprocessor/internal/metrics/metrics.go
+++ b/processor/deltatocumulativeprocessor/internal/metrics/metrics.go
@@ -47,7 +47,7 @@ func (m Metric) AggregationTemporality() pmetric.AggregationTemporality {
 	return pmetric.AggregationTemporalityUnspecified
 }
 
-func (m Metric) Typed() any {
+func (m Metric) Typed() Any {
 	//exhaustive:enforce
 	switch m.Type() {
 	case pmetric.MetricTypeSum:
@@ -62,4 +62,50 @@ func (m Metric) Typed() any {
 		return Summary(m)
 	}
 	panic("unreachable")
+}
+
+var (
+	_ Any = Sum{}
+	_ Any = Gauge{}
+	_ Any = ExpHistogram{}
+	_ Any = Histogram{}
+	_ Any = Summary{}
+)
+
+type Any interface {
+	Len() int
+	Ident() identity.Metric
+
+	SetAggregationTemporality(pmetric.AggregationTemporality)
+}
+
+func (m Metric) Filter(ok func(id identity.Stream, dp any) bool) {
+	mid := m.Ident()
+	switch m.Type() {
+	case pmetric.MetricTypeSum:
+		m.Sum().DataPoints().RemoveIf(func(dp pmetric.NumberDataPoint) bool {
+			id := identity.OfStream(mid, dp)
+			return !ok(id, dp)
+		})
+	case pmetric.MetricTypeGauge:
+		m.Gauge().DataPoints().RemoveIf(func(dp pmetric.NumberDataPoint) bool {
+			id := identity.OfStream(mid, dp)
+			return !ok(id, dp)
+		})
+	case pmetric.MetricTypeHistogram:
+		m.Histogram().DataPoints().RemoveIf(func(dp pmetric.HistogramDataPoint) bool {
+			id := identity.OfStream(mid, dp)
+			return !ok(id, dp)
+		})
+	case pmetric.MetricTypeExponentialHistogram:
+		m.ExponentialHistogram().DataPoints().RemoveIf(func(dp pmetric.ExponentialHistogramDataPoint) bool {
+			id := identity.OfStream(mid, dp)
+			return !ok(id, dp)
+		})
+	case pmetric.MetricTypeSummary:
+		m.Summary().DataPoints().RemoveIf(func(dp pmetric.SummaryDataPoint) bool {
+			id := identity.OfStream(mid, dp)
+			return !ok(id, dp)
+		})
+	}
 }