--- a/processor/deltatocumulativeprocessor/internal/testing/sdktest/metrics.go
+++ b/processor/deltatocumulativeprocessor/internal/testing/sdktest/metrics.go
@@ -123,3 +123,14 @@ func (attr attributes) Into() attribute.Set {
 //
 // Temporality is optional and defaults to [sdk.CumulativeTemporality]
 type Format = []byte
+
+func Expect(metrics map[string]Metric) Spec {
+	for name, m := range metrics {
+		m.Name = name
+		if m.Temporality == 0 {
+			m.Temporality = sdk.CumulativeTemporality
+		}
+		metrics[name] = m
+	}
+	return Spec(metrics)
+}