--- a/processor/deltatocumulativeprocessor/linear.go
+++ b/processor/deltatocumulativeprocessor/linear.go
@@ -5,7 +5,6 @@ package deltatocumulativeprocessor // import "github.com/open-telemetry/opentele
 
 import (
 	"context"
-	"errors"
 	"sync"
 	"time"
 
@@ -16,22 +15,19 @@ import (
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/exp/metrics/identity"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/exp/metrics/staleness"
-	exp "github.com/open-telemetry/opentelemetry-collector-contrib/internal/exp/metrics/streams"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/deltatocumulativeprocessor/internal/data"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/deltatocumulativeprocessor/internal/delta"
 	telemetry "github.com/open-telemetry/opentelemetry-collector-contrib/processor/deltatocumulativeprocessor/internal/lineartelemetry"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/deltatocumulativeprocessor/internal/metrics"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/deltatocumulativeprocessor/internal/streams"
 )
 
 var _ processor.Metrics = (*Linear)(nil)
 
 type Linear struct {
-	next processor.Metrics
+	next consumer.Metrics
 	cfg  Config
 
-	state state
-	mtx   sync.Mutex
+	last state
+	mtx  sync.Mutex
 
 	ctx    context.Context
 	cancel context.CancelFunc
@@ -40,16 +36,16 @@ type Linear struct {
 	tel   telemetry.Metrics
 }
 
-func newLinear(cfg *Config, tel telemetry.Metrics, next processor.Metrics) *Linear {
+func newLinear(cfg *Config, tel telemetry.Metrics, next consumer.Metrics) *Linear {
 	ctx, cancel := context.WithCancel(context.Background())
 
 	proc := Linear{
 		next: next,
 		cfg:  *cfg,
-		state: state{
-			nums: make(exp.HashMap[data.Number]),
-			hist: make(exp.HashMap[data.Histogram]),
-			expo: make(exp.HashMap[data.ExpHistogram]),
+		last: state{
+			nums: make(map[identity.Stream]pmetric.NumberDataPoint),
+			hist: make(map[identity.Stream]pmetric.HistogramDataPoint),
+			expo: make(map[identity.Stream]pmetric.ExponentialHistogramDataPoint),
 		},
 		ctx:    ctx,
 		cancel: cancel,
@@ -58,7 +54,7 @@ func newLinear(cfg *Config, tel telemetry.Metrics, next processor.Metrics) *Line
 		tel:   tel,
 	}
 
-	tel.WithTracked(proc.state.Len)
+	tel.WithTracked(proc.last.Len)
 	cfg.Metrics(tel)
 
 	return &proc
@@ -75,68 +71,68 @@ func (p *Linear) ConsumeMetrics(ctx context.Context, md pmetric.Metrics) error {
 		drop = false
 	)
 
-	// possible errors encountered while aggregating.
-	// errors.Join-ed []streams.Error
-	var errs error
-
 	metrics.Filter(md, func(m metrics.Metric) bool {
 		if m.AggregationTemporality() != pmetric.AggregationTemporalityDelta {
 			return keep
 		}
 
-		// NOTE: to make review and migration easier, below only does sums for now.
-		// all other datatypes are handled by older code, which is called after this.
-		//
-		// TODO: implement other datatypes here
-		if m.Type() != pmetric.MetricTypeSum {
-			return keep
-		}
+		// aggregate the datapoints.
+		// using filter here, as the pmetric.*DataPoint are reference types so
+		// we can modify them using their "value".
+		m.Filter(func(id identity.Stream, dp any) bool {
+			// count the processed datatype.
+			// uses whatever value of attrs has at return-time
+			var attrs telemetry.Attributes
+			defer func() { p.tel.Datapoints().Inc(ctx, attrs...) }()
 
-		sum := metrics.Sum(m)
-		state := p.state.nums
-
-		// apply fn to each dp in stream. if fn's err != nil, dp is removed from stream
-		err := streams.Apply(sum, func(id identity.Stream, dp data.Number) (data.Number, error) {
-			acc, ok := state.Load(id)
-			// if at stream limit and stream not seen before, reject
-			if !ok && p.state.Len() >= p.cfg.MaxStreams {
-				p.tel.Datapoints().Inc(ctx, telemetry.Error("limit"))
-				return dp, streams.Drop
+			// if stream new and state capacity reached, reject
+			exist := p.last.Has(id)
+			if !exist && p.last.Len() >= p.cfg.MaxStreams {
+				attrs.Set(telemetry.Error("limit"))
+				return drop
 			}
 
-			// stream is alive, update stale tracker
+			// stream is ok and active, update stale tracker
 			p.stale.Refresh(now, id)
 
-			acc, err := func() (data.Number, error) {
-				if !ok {
-					// new stream: there is no existing aggregation, so start new with current dp
-					return dp.Clone(), nil
-				}
-				// tracked stream: add incoming delta dp to existing cumulative aggregation
-				return acc, delta.AccumulateInto(acc, dp)
-			}()
-			// aggregation failed, record as metric and drop datapoint
-			if err != nil {
-				p.tel.Datapoints().Inc(ctx, telemetry.Cause(err))
-				return acc, streams.Drop
+			// this is the first sample of the stream. there is nothing to
+			// aggregate with, so clone this value into the state and done
+			if !exist {
+				p.last.BeginWith(id, dp)
+				return keep
 			}
 
-			// store aggregated result in state and return
-			p.tel.Datapoints().Inc(ctx)
-			_ = state.Store(id, acc)
-			return acc, nil
+			// aggregate with state from previous requests.
+			// delta.AccumulateInto(state, dp) stores result in `state`.
+			// this is then copied into `dp` (the value passed onto the pipeline)
+			var err error
+			switch dp := dp.(type) {
+			case pmetric.NumberDataPoint:
+				state := p.last.nums[id]
+				err = delta.AccumulateInto(state, dp)
+				state.CopyTo(dp)
+			case pmetric.HistogramDataPoint:
+				state := p.last.hist[id]
+				err = delta.AccumulateInto(state, dp)
+				state.CopyTo(dp)
+			case pmetric.ExponentialHistogramDataPoint:
+				state := p.last.expo[id]
+				err = delta.AccumulateInto(state, dp)
+				state.CopyTo(dp)
+			}
+			if err != nil {
+				attrs.Set(telemetry.Cause(err))
+				return drop
+			}
+
+			return keep
 		})
-		errs = errors.Join(errs, err)
 
-		// all remaining datapoints are cumulative
-		sum.SetAggregationTemporality(pmetric.AggregationTemporalityCumulative)
-
-		// if no datapoints remain, drop now-empty metric
-		return sum.Len() > 0
+		// all remaining datapoints of this metric are now cumulative
+		m.Typed().SetAggregationTemporality(pmetric.AggregationTemporalityCumulative)
+		// if no datapoints remain, drop empty metric
+		return m.Typed().Len() > 0
 	})
-	if errs != nil {
-		return errs
-	}
 
 	// no need to continue pipeline if we dropped all metrics
 	if md.MetricCount() == 0 {
@@ -159,7 +155,7 @@ func (p *Linear) Start(_ context.Context, _ component.Host) error {
 					p.mtx.Lock()
 					stale := p.stale.Collect(p.cfg.MaxStale)
 					for _, id := range stale {
-						p.state.Delete(id)
+						p.last.Delete(id)
 					}
 					p.mtx.Unlock()
 				}
@@ -179,33 +175,40 @@ func (p *Linear) Capabilities() consumer.Capabilities {
 	return consumer.Capabilities{MutatesData: true}
 }
 
-type Metric[T data.Point[T]] interface {
-	metrics.Filterable[T]
-	SetAggregationTemporality(pmetric.AggregationTemporality)
-}
-
 // state keeps a cumulative value, aggregated over time, per stream
 type state struct {
-	nums streams.Map[data.Number]
-
-	// future use
-	hist streams.Map[data.Histogram]
-	expo streams.Map[data.ExpHistogram]
+	nums map[identity.Stream]pmetric.NumberDataPoint
+	hist map[identity.Stream]pmetric.HistogramDataPoint
+	expo map[identity.Stream]pmetric.ExponentialHistogramDataPoint
 }
 
 func (m state) Len() int {
-	return m.nums.Len() + m.hist.Len() + m.expo.Len()
+	return len(m.nums) + len(m.hist) + len(m.expo)
 }
 
 func (m state) Has(id identity.Stream) bool {
-	_, nok := m.nums.Load(id)
-	_, hok := m.hist.Load(id)
-	_, eok := m.expo.Load(id)
+	_, nok := m.nums[id]
+	_, hok := m.hist[id]
+	_, eok := m.expo[id]
 	return nok || hok || eok
 }
 
 func (m state) Delete(id identity.Stream) {
-	m.nums.Delete(id)
-	m.hist.Delete(id)
-	m.expo.Delete(id)
+	delete(m.nums, id)
+	delete(m.hist, id)
+	delete(m.expo, id)
+}
+
+func (m state) BeginWith(id identity.Stream, dp any) {
+	switch dp := dp.(type) {
+	case pmetric.NumberDataPoint:
+		m.nums[id] = pmetric.NewNumberDataPoint()
+		dp.CopyTo(m.nums[id])
+	case pmetric.HistogramDataPoint:
+		m.hist[id] = pmetric.NewHistogramDataPoint()
+		dp.CopyTo(m.hist[id])
+	case pmetric.ExponentialHistogramDataPoint:
+		m.expo[id] = pmetric.NewExponentialHistogramDataPoint()
+		dp.CopyTo(m.expo[id])
+	}
 }