@sh0rez 2024-11-28
@ArthurSens while the benchmark is split by datatype, I only ran for sums as the processor part (stream tracking, etc) is the same across all datatypes. speed differences only come by add implementation